+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/compose/TooltipIconButton.kt
@@ -11,7 +11,9 @@ import androidx.compose.material3.TooltipBox
 import androidx.compose.material3.TooltipDefaults
 import androidx.compose.material3.rememberTooltipState
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.painter.Painter
 import org.jetbrains.compose.resources.DrawableResource
 import org.jetbrains.compose.resources.painterResource
 
@@ -39,6 +41,31 @@ fun TooltipIconButton(
                 painter = painterResource(icon),
                 contentDescription = null // tooltip provides description
             )
+        }
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TooltipIconButton(
+    tooltip:String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true,
+    content: @Composable (() -> Unit)
+) {
+    val tooltipState = rememberTooltipState()
+    TooltipBox(
+        positionProvider = TooltipDefaults.rememberTooltipPositionProvider(TooltipAnchorPosition.Above),
+        state = tooltipState,
+        tooltip = {
+            PlainTooltip {
+                Text(tooltip)
+            }
+        },
+    ) {
+        IconButton(onClick,modifier,enabled) {
+            content()
         }
     }
 }