+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/key/component/AIAgentListComponent.kt
@@ -1,9 +1,10 @@
 package ai.mcpdirect.studio.app.key.component
 
+import ai.mcpdirect.mcpdirectstudioapp.JSON
 import ai.mcpdirect.studio.app.compose.StudioActionBar
 import ai.mcpdirect.studio.app.compose.StudioListItem
 import ai.mcpdirect.studio.app.tips.AIAgent
-import ai.mcpdirect.studio.app.tips.aiAgents
+import ai.mcpdirect.studio.app.tips.aiAgentIntegrationGuide
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.lazy.LazyColumn
@@ -23,18 +24,24 @@ fun AIAgentListComponent(
     modifier: Modifier,
     onAIAgentChange:(aiAgent:AIAgent)->Unit
 ){
-    var aiAgent by remember { mutableStateOf(aiAgents[0]) }
+    var aiAgents by remember {mutableStateOf<List<AIAgent>>(emptyList())}
+    var aiAgent by remember { mutableStateOf<AIAgent?>(null) }
     LaunchedEffect(Unit) {
-        onAIAgentChange(aiAgent)
+        aiAgents = JSON.decodeFromString(aiAgentIntegrationGuide)
+        if(aiAgents.isNotEmpty()){
+            aiAgent = aiAgents[0]
+            onAIAgentChange(aiAgent!!)
+        }
+
     }
     Column(modifier) {
         StudioActionBar("AI Agents")
         HorizontalDivider()
         LazyColumn(Modifier.weight(1f)) {
             items(aiAgents) {
-                if(aiAgent==null){
-                    aiAgent = it
-                }
+//                if(aiAgent==null){
+//                    aiAgent = it
+//                }
                 StudioListItem(
                     modifier = Modifier.clickable {
                         aiAgent = it